+++ src/components/Common/InputErrorMessage.tsx
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+const InputErrorMessage = styled.p`
+  color: var(--color-error);
+  font-size: 12px;
+  line-height: 14px;
+  height: 14px;
+  margin: 0;
+`;
+
+export default InputErrorMessage;
